Fix e2e test stability (#761)
This is a continuation of issue #750. In summary, we're encountering
intermittent test failures. Upon deleting the pod in admintools
deployments, we observed that the Vertica process shows up in the `ps`
output. A new pod should not have Vertica already running because the
operator should be triggering that. Although this doesn't impact the
operator's functionality, it disrupts end-to-end tests for upgrades, as
they rely on the operator's sequential actions.

I encountered another instance of this issue while investigating with
the debug code introduced in #750. It's evident from the `ps` output
that the Vertica process belongs to a previous instance of the pod. One
clear indication is that the host IP provided to the Vertica process
doesn't match the pod's IP. Therefore, I've implemented this check to
address the test failure.