Remove minio/console pull
The minio/console repository is no longer in docker hub. We don't use
the console for anything. It was pulled because a pod is created that
usese it when deploy the minio operator. But we have no use for it.
Note, the k8s CI will still attempt to pull the image. But it will get
ErrImagePull and it won't affect any tests because we don't rely on it.