Decouple subcluster name from its statefulset name (#780)
This adds a new subcluster annotation to control the statefulset name.
The new annotation is: `vertica.com/statefulset-name-override`. This is
needed for replicated upgrade when the sandbox gets promoted to the main
cluster. We will want to rename the sandbox subcluster to match the
original names. However, the statefulset name always included the
subcluster name. This new annotation will allow us to rename the
subcluster while keeping the statefulset name constant. This ensures we
can do a rename without needing to rebuild the statefulset.