Add offline upgrade reconciler to sandbox controller (#785)
This will allow us to upgrade the subclusters in a sandbox. There is a
new image field in the VerticaDB for the sandbox. When that is changed,
it will cause the sandbox controller to drive the upgrade. The vdb
controller will check if a sandbox needs an upgrade and will wake up the
sandbox controller by setting the vdb resource version in the configmap
annotations.
You can follow a sandbox upgrade progress through the upgradestate field
in the sandbox status.