Add support for multiple subclusters in a sandbox (#799)
This adds support for multiple subclusters in a single sandbox. In order
to denote which subclusters are primary in the sandbox, a new subcluster
type is added called 'sandboxprimary'. New webhook rules have been added
to account for the new subcluster type.

The new type will get added to the first subcluster in a sandbox. We now
add subclusters to a sandbox in the order they are defined in the vdb
spec. This ensures that we pick a deterministic subcluster as the first
one in a sandbox.

When adding a subcluster to an existing sandbox, we need to pass in a
host from the main cluster and the sandbox too. This reason this is
needed is that we need to replicate the new sandbox state in both the
main cluster and in the sandbox. So, the vclusterOps API will update
both clusters.