Call sandbox reconciler during replicated upgrade (#807)
This builds out the replicated upgrade logic more. With the sandbox
controller and the replicator controller both implemented, the upgrade
logic can now depend on them. A few changes were made while verifying
the new controllers. We can now get to the part where we are waiting for
promotion of the sandbox to the main cluster. All traffic is routed to
the sandbox, which have been upgraded to the new server version.