Decouple sandbox image from vdb.spec image (#812)
When setting a sandbox, if the image is omitted, the operator will
internally assume spec.image as the sandbox's image. The issue is that
changing spec.image could trigger sandbox upgrade too. Now if the image
is omitted, there is a new reconciler that will explicitly set the
sandbox image in vdb to spec.image value.

---------

Signed-off-by: roypaulin <roy.paulin@microfocus.com>