Applied new option --hosts-in-sandbox of vclusterOps start_db on the operator (#813)
When restarting a pod in a database with a sandbox, it could fail when
the IP of the pod changes. This is a bug in old vclusterOps code. We
were trying to read node IPs from the catalog, but sandbox node IP
change cannot be reflected in the main cluster catalog, and vice versa.
This PR applied the new vclusterOps version and passed a new option to
start_db to fix this issue. With this option, we will read main cluster
node IPs from main cluster catalog and read sandbox node IPs from
sandbox catalog.