Fixed bugs in sandbox and unsandbox (#820)
This PR applies the latest vclusterOps and passed more node information
to vclusterOps for fixing the bugs in sandbox and unsandbox. The bugs
can be reproduced when we kill a pod in a subcluster before we sandbox
it and kill a pod in a sandbox before we unsandbox it. The issue is
cause by the stale in the catalog. The vcluster re-ip only can modify
the nodes' IPs on either main cluster or the sandbox. The new
vclusterOps will do one more re-ip before sandboxing/unsandboxing.