Fixed intermittent failure in online-upgrade-kill-transient test (#825)
This PR applied new vclusterOps library to fix intermittent failure in
online-upgrade-kill-transient test. That intermittent failure is caused
by IP inconsistency between the transient node and the catalog. In the
new vclusterOps library, we won't check the version of the transient
node in start_node, and we will do re-ip before removing the transient
node.