Fixed Online Upgrade Error (#896)
When we run an online upgrade from Version 24.2.x to 24.3.x, the
operator will fail due to pod facts collection error. The reason is that
secondary subclusters still use the old Vertica Version 24.2.x when
primary subclusters finished the upgrade. As a result, the operator will
call vclusterOps API to collect node details on secondary subclusters,
but secondary subclusters doesn't have that API due to a low Vertica
version.

This PR saved the old Vertica version prior to the upgrade, and used
that version to decide if the operator should call vclusterOps API to
collect node details during the upgrade.