Fixed online upgrade halting issue (#898)
This PR fixed online upgrade halting issue after sandbox restart. We use
vclusterOps to collect node details in online upgrade rather than vsql.
This change will remove active connections on the subclusters so
subclusters doesn't need to wait for user connections to be paused. As a
result, it will fix the halting issue.