Tighten the check for expired sandbox config map (#907)
There is an ND error during the online upgrade. After a successful
sandboxing, the config map is deleted by the sandbox controller which is
not desired. There is a delay for sandbox controller to get the latest
sandbox status. I made the expired sandbox check more strict in this PR.
You can download the error logs from Jira VER-96105.