config kind cluster to increase maximum log file size and event-ttl (#912)
A larger operator log size and longer events duration would be useful
for users who want to debug issues on a local Kind cluster. This PR will
increase the maximum log size from 10Mi to 100Mi and the duration of
events from 1 hour to 24 hours.

# Patch in kubelet to increase maximum log file size
On a newly deployed Kubernetes cluster, the pod's log is located under
the host /var/log/pods/<pods-name> directory with a maximum log file
size of 10Mi (openshift cluster is 50Mi).
As our verticadb-operator pod's log grows fast as time passes by,
especially with online upgrade operations, it's recommended to configure
the kubelet argument containerLogMaxSize to increase the maximum log
file size to 100Mi.

# Patch in apiserver to increase events ttl duration
By default, the k8s cluster clears the events after 1 hour and is not
persistent on any disk. The direct way is to add extra args --event-ttl
for apiserver to increase the events time-to-live (TTL) duration.