Fix admintools read-only upgrade bug from a version < v12.0.4 to a version > v12.0.4 (#946)
During podfact collection, we add the sandbox state to the query to get
node details only for version >= v12.0.4. The issue is that we get the
version from the `version` annotation and in the middle of read-only
online upgrade it will contain the new version. Therefore, The `select
... from nodes;` will fail for subclusters that still run the old image,
if that image is < v12.0.4, because it contains the sandbox state which
is not supported. That is why the daily run fails(base_image<v12.0.4 and
target_image>v12.0.4).
To fix it, the operator will get the version from the `previous-version`
annotation if read-only online upgrade is in progress.