Prevent skipping some tests (#952)
The reason why normal ci tests were oddly more stable than daily build
tests was because we were skipping a lot of them.
We rely on the vertica-version label in the image to skip some tests in
case the vertica version is lower than the min required version by the
test. The issue is that label is empty for `latest-test-master` which
leads to all the tests using that image to be skipped. We need to set
that label properly for that image too but in the meantime, we are
adding this work-around to not skip some tests.