Add support for stop sandbox (#955)
This is the first step to support stop sandbox for realdb. Normally, if
a pod is down the operator will try to restart it. However, for the
realdb use case we want to be able to stop a standbox and to prevent the
operator from restarting it until we decide otherwise.
This adds the logic to stop_db a sandbox when
`spec.sandboxes[].shutdown` is true, a new reconciler has been added to
the vdb controller to wake up the sandbox controller. As long as that
field is true, the operator will ignore the sandbox's pods. As soon as
it is set to false, the operator will restart the sandbox.

Scaling down the sts and adding webhook rules will be done on a
follow-up ticket.