Delete a subcluster pods after it is shut down (#968)
After a subcluster is "shut down"(shutdown set to true), its pods will
stay down and cannot get restarted by the operator. We go a step further
and delete the pods by scaling the statefulset to zero. This is useful
for realdb as we do not want to keep down pods around and the EKS
instances will also be shut down.
When shutdown is set back to false, the statefulset will be scale up to
its original size(subcluster size in spec).