Fix sandbox creation failure when the sandbox name contains underscore (#974)
The sandbox name is used as a suffix in the configMap name. The issue is
that k8s objects do not allow underscore in their name so when there is
an underscore in the sandbox name we end up with one in the configMap
which leads to failure during its creation. Given underscore is a valid
character for a sandbox name, we are going to replace it in the
configMap name with hyphen.

Additionally, we added a check of the sandbox name to make sure it
matches rfc 1123 regex.