Prevent proxy deployment update (#1002)
When creating a deployment, a replicaset is created for each replica.
When the update strategy is `RollingUpdate`, each time there is a change
in the deployment spec, an extra pod will be created with the update and
the old one will be deleted once the new one is ready. So there is a
timing where you may see twice the number of pods with respect to what
you expect.
During testing, QA found that each time we create a proxy deployment, at
first we had twice the number of required pods before getting the
correct number after a few seconds. That is because we updated the
deployment by adding the server version annotation when it is available
in the vdb. That annotation does't serve any purpose so I just removed
to avoid updating the deployment.