Fix sandbox-upgrade-happy-path (#1006)
A bug was found in sandbox-upgrade-happy-path test case. The sandbox is
upgraded and then removed (unsandboxed). After unsandboxing, we restart
all the pods of the former sandbox. The issue is that the subcluster
that joins the main cluster is not running the same image, so we get
stuck for a while in startnode(HTTPSPollNodeStateIndirectOp) until the
timeout(1200s). After timeout expires, it will requeue and run
`UnsandboxImageVersionReconciler` which will recreate the pods with the
correct image and next the next startnode attempt will succeed. The
solution consists in calling `UnsandboxImageVersionReconciler` just
before the restart reconciler.