Fix admintools e2e tests (#1026)
The failures are due to a recent update to the GitHub
runner:https://github.com/actions/runner-images/releases/tag/ubuntu24%2F20250105.1.
This update requires all sudo commands to prompt for a non-empty
password(dbadmin doesn't have a password). I am unable to determine the
root cause and resolve the issue, the current workaround is to use an
earlier version of the Ubuntu runner.

I will create a follow-up Jira ticket to address the sudo issue. At
present, the issue only affects admintools, so it is not critical.