Add threshold for scale down (#1038)
This adds a new field `scaleDownThreshold` to the metric's definition.
It will be used to control scale down i.e scale down will happen only
when the metric's value is lower than this new field's value.
As the hpa does not natively support 2 thresholds, this works through a
hack: we know that scaling cannot happen if the currentReplicas is equal
to the minReplicas so the operator sets the hpa minReplicas to the
VerticaAutoscaler's targetSize as long as the metric's value is greater
than scaleDownThreshold thus preventing scale down. The moment the value
is lower than scaleDownThreshold, the operator update the hpa
minReplicas to its correct value (from vas) which will trigger scale
down.
This will not work as expected with the scale-down stabilization window
as the new threshold is not part of the hpa. So, it will be recommended
to either set a stabilization window or set scaleDownThreshold for a
stable scaling down experience

NB: Using the scale-down stabilization window is also a very good way to
have a stable scaling down process. A stabilization window long enough
will assure the metric has been lower long enough to safely scale down.