Add routing label to sandbox pods (#1045)
When a sandbox pod is deleted, the operator does not put back the
routing label. This adds the ClientRoutingLabelReconciler in the sandbox
controller. It also adds an annotation in the configMap to sometimes
disable routing, like in the case of online upgrade where routing to the
sandbox must not be allowed until replication is done.