add type service for vertica_queued_requests_total (#1043)
The type service is missing for vertica_queued_requests_total in
Prometheus adapter configurations. This caused the VerticaAutoscaler
sample configuration vertica_queued_requests_count to not work.

This PR added the type service to the metric and also added missing test
files related to the Prometheus adapter setup and cleanup.