Support autoscaling with KEDA (#1042)
KEDA (Kubernetes-based Event Driven Autoscaling) is an open-source that
allows for fine-grained autoscaling (including to/from zero) for event
driven Kubernetes workloads. This PR integrates KEDA's ScaledObject to
the VerticaAutoscaler CR. Now users can pick which autoscaler to use
(scaledObject/hpa).
API changes were made now the customAutoscaler field has 3 subfields:
type, hpa and scaledObject. When type is "HPA", hpa must be set with the
metric's definition. The metric can either be cpu/mem or a prometheus
metric.
A new test case was added to scale based on a scaledObject.