Create v1 vertica autoscaler (#1054)
With the changes we are now making on the VerticaAutoscaler, it is time
we release a v1 version. We will keep both v1 and v1beta1 around and
will slowly deprecate v1beta1.

As part of this work, for any existing field that would need an update,
we will do it in v1 and use a conversion webhook to maintain
consistency. We also need to change the status conditions field in v1 in
order to use the structures from metav1.

---------

Co-authored-by: roypaulin <rnguetsopken@opentext.com>