Add method for triggerAuthentication (#1066)
So far, we have been running a prometheus instance that does not require
any authentication but custom will have some kind of authentication in
place.

For Keda ScaledObject: we need to add an option to pass a secret that
contains prometheus creds and use it to successfully connect to
prometheus. It means the operator will create a KEDA
triggerAuthentication object

We will add two new fields authModes and unsafeSsl in prometheusSpec
struct.
authModes with four types: 'basic', 'bearer', 'tls', 'custom',
'tls,basic'.
unsafeSsl: default with false

Ref: https://keda.sh/docs/2.16/scalers/prometheus/

---------

Co-authored-by: roypaulin <rnguetsopken@opentext.com>