Jc4244441 Replace password with certificate  (#1056)
In my previously merged PR, support for using env var to pass secret
name is added for NMA.
Here are the major new changes added:
1 The operator will use password to create db, execute DDLs to configure
TLS.
Only when DDLs are executed properly will the operator starts to use
certificate.
2 The decision about whether to use password or certificate is based on
the following factors:
    -     vdb version
    -     tls annotation
    -     whether DB is initialized
3 I realize it is hard to let functions in vadmin/ package to share the
data as it is a
multithreading env. To make it easier to set/pass a flag, I added a new
data structure.
VdbContext. It is a thread safe map that supports multiple databases. We
can use it
to store bool or string values in it. This data structure will also be
used when we rotate
    certificate.

Use of "UseTLSCert" flag
This vdb-scoped global flag indicates whether we should use TLS instead
of password.
It will be set up before a reconcile cycle starts, based on 
    -     vdb version
    -     tls annotation
    -     whether DB is initialized
We use it as it makes it a lot easier to pass parameter from other
packages to vadmin/

---------

Co-authored-by: roypaulin <rnguetsopken@opentext.com>