Add timeout to drain before remove node/subcluster (#1175)
The current drain just requeue as long as there is at least one
connection to a pending delete node.
This adds a timeout to allow users to close the active sessions on
pending pods once the timeout has expired.
2 new annotations:
- vertica.com/active-connections-drain-seconds: the timeout
- vertica.com/drain-start-time: the time at which draining started

We don't use a sleep because it is very inefficient, we instead requeue
after 1s until the timeout expires, to not block a thread.