feat(annotations): add vertica.com/disable-extra-paths-auto-mount (#1… (#1178)
…177)

This complements the disable-depot-volume-management annotation.

When a database is revived, the extraPaths annotation get automatically
populated from the metadata. This has negative effects when volumes are
managed manually such as when they're self-managed. In our case, the
temp volume is on a local fast SSD, and our config ("local-data" in
Vertica parlance) is on a RBD to allow for rescheduling. We don't want
temp storage on RBD, so we don't want to get our TEMP storage location
force-mounted as the RBD.

Using this annotation allows to skip the extraPaths auto-mounting thus
disabling this detrimental feature in an explicit way.

# How to use 

Set the following annotation in your VerticaDB object:

`"vertica.com/disable-extra-paths-auto-mount": "true"`

---------

Co-authored-by: Matcha <3207830+matchaxnb@users.noreply.github.com>
Co-authored-by: Matcha (from Criteo) <m.desoutter@criteo.com>