Set max number of restore points for archive when saving restore point (#1209)
We can store a restore point by setting
"status.conditions.SaveRestorePointNeeded" to true. This will
automatically create an archive, if it does not exist.

Previously, this archive would have max restore points set to 0, meaning
we can save an unlimited number of restore points for this archive.

Now, the maximum number of restore points can be controlled by setting
"spec.restorePoints.numRestorePoints". When this maximum is exceeded, we
will overwrite the oldest restore point. Default is 0, meaning that
number of restore points is still unlimited.

NOTE: this only applies when creating an archive. If you have an
existing archive, it will not update the max number of restore points on
it. This scenario can be handled in a separate PR (as it requires
vcluster API changes).

(New PR; made a mistake doing a rebase)