Support aws secret version id (#1216)
In aws secrets manager, a secret can have multiple versions that show
how it changed over time. Previously, users had to create a new secret
in aws secrets manager, each time they want to rotate the old one.
Now we allow users to pass a specific version of a secret in aws secrets
manager.
The new format is: awssm://<secret-arn>@<version-id> where `version-id`
is the specific version of the secret you want to use.
This applies to all secrets used in the CRs

Another change is the choice of caching the the current secret names in
the status instead of the annotations. For that a new field `secretRefs`
has been added to the status. It contains the names of the secrets
currently in use.
Ex: spec.nmaTLSSecret = "sec1", spec.passwordSecret = "sec2" The status
will look like this:
```
secretRefs:
- name: sec1
   type: NMATLSType
- name: sec2
   type: PASSWORDType
```
As you can see, this will be very easy to extend and we don't have to
worry about users accidently changing them.