Allow non-default ports for Vertica services (#1218)
Currently, we hardcode port 8443 for HTTPS services and port 5433 for
the client. However, there is a request to be able to customize these
port. In the customer's case, they have another product running on 8443,
which they cannot shut down, so they want Vertica to use a different
port. Also, it would be nice if we could set different subclusters to
use different HTTPS ports.

This will add four new optional fields to the VDB:

1. spec.serviceHTTPSPort: port for ALL HTTPS services; default is 8443
2. spec.serviceClientPort: client port for all services; default is 5433
3. spec.subclusters[].serviceHTTPSPort: port of the HTTPS service of an
individual subcluster; this will take precedence over
spec.serviceHTTPSPort
4. spec.subclusters[].serviceClientPort: client port of the service of
an individual subcluster; this will take precedence over
spec.serviceHTTPSPort