Fix sandbox_shutdown (#1262)
In a previous PR, I decided to call sandboxshutdown _reconciler early in
the reconcile loop, to prevent some failures to affect some sandbox ops.
The issue is that it might be possible that (ex: realdb) someone is
adding a new sandbox +subclusters and that will fail sandboxshutdown
_reconciler. The 2nd sandboxshutdown _reconciler does not have this
issue because it happens towards the end when all critical reconcilers
have been run.
To fix this, I added a field IgnoreError that is true for the 1st
sandboxshutdown _reconciler. When true, if a subcluster is not yet in
the status, we just skip that sc knowing a later reconciler will add it
to the db.