Add new tool to generate tls cert (#1278)
This adds a new binary that can be used to generated a valid tls secret
for nma/https and client-server. This can be handy when you do not want
to rely on the operator to automatically generate a secret as that
secret will get deleted with the vdb. This will also make testing easier
as we won't have to copy an autogenerated secret and won't have to
manually create a tls secret (which might be invalid). We also do not
have to store secret yaml files for the e2e tests.

This tool can be improved and even exposed to users and QA.