Modify annotations for TLS authentication (#1292)
Currently, to enable TLS authentication, we must set two different
annotations:

- enable-tls-certs-rotation must be set to True
- mount-nma-certs must be set to False

This is inconvenient. We should be able to set a single annotation to
enable/disable TLS. So, we have renamed:

`enable-tls-certs-rotation`

To

`enable-tls-auth`

And this will now be the only annotation required for TLS;
mount-nma-certs will no longer impact TLS.

We also have a requirement that mount-nma-certs be set to False by
default, for security reasons.