Fix bug in obj_reconciler (#1311)
The issue is "reconcileTLSSecrets()" in obj_reconciler.go. That function
must be used for k8s secrets only. When the secret is in awssm, we would
try to get it using k8s api, which of course will fail. Now we first
check if the tls secret is in k8s and skip it if that's not the case