Add timeout to drain during read-only online upgrade (#1318)
During read-only online upgrade, we do a shutdown of all primary
subclusters and all secondary subclusters. Before completing this
shutdown, we re-route all client traffic and then do a drain of all
existing connections. Unlike in drainnode_reconciler (which will
eventually timeout and kill the sessions), this drain will continue
infinitely.

We would like to add a timeout to this drain logic. If it drains for
more than a certain amount of time, we will auto-kill the sessions and
proceed with the the shutdown and upgrade.

For this, we will re-use much of the logic from drainnode_reconciler. So
the timeout can be configured via annotation
vertica.com/active-connections-drain-seconds, defaulted to 60. However,
in this case, we are draining subcluster by subcluster, so we cannot
re-use the entire reconciler.

FUTURE WORK: Currently, it will drain subclusters sequentially. However,
in the future, it would be nice to include multiple subclusters in a
single drain (for example, all primaries and then all secondaries). This
would speed up the process for customers with many subclusters and many
active connections.