Fixed operator upgrade issue (#1331)
When we upgrade operator to 25.3, we will update the statefulSet due to
the change of NMA env vars and mounted volume. This PR added check for
nma-mount-cert and it can keep the mounted volume and only use mounted
envs after the upgrade.

Moreover, I removed operator-version annotation from pod template so
that the future upgrade won't trigger rolling restart of the database.