Create HTTPS/ClientServer TLS rollback skeleton (#1335)
This lays the foundations for rollback after tls cert rotation failure. 
The failure scenarios are the following (simplified summary):
- https/clientserver cert rotation fails but before any change was made
in the tls config. We are still using the old cert so we just need to
undo the secret name change. (no-op on nma)
- https/clientserver cert rotation fails, we altered the tls config but
polling the new cert health failed (it is unlikely): We need to rollback
the rotation by re-running cert rotation with reverse params; to get
back to the old cert. (no-op on nma)
- https/clientserver cert rotation succeeded but nma cert rotation
failed: we need to roll back both nma and https. First nma, we need to
restart nma containers with the old cert. Once nma is back running with
old cert, we re-run cert rotation with reverse params;

I added a new reconciler that will incapsulate the rollback logic. In
the cert rotation reconcilers, we will set a status condition with a
specific "reason". The rollback reconciler will use that condition to
decide how far it needs to go to undo what has been done.

This is only the skeleton and currently does not do any rollback. 

This also changes how we update the NMA ConfigMap. Previously, this was
done before https/clientserver cert rotation, and the cert rotation
reconcilers would verify that it had been updated. Now, it is done after
the rotations and, if either rotation has failed, it will be skipped;
this ensures we are always using the previous good certs after failure.
Also, we have moved the checking of the ConfigMap to the NMA cert
reconciler.

---------

Co-authored-by: roypaulin <rnguetsopken@opentext.com>